requirements.txt
asgiref==3.5.2
boto3==1.23.5
botocore==1.26.5
cffi==1.15.0
click==8.1.3
cmake==3.22.4
cryptography==37.0.2
Django==4.0.3
django-crispy-forms==1.14.0
django-filter==21.1
django-multiselectfield==0.1.12
django-rest-knox==4.2.0
djangorestframework==3.13.1
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
jwt==1.3.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
mysqlclient==2.1.0
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.1
protobuf==3.20.1
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
